Move invites models to models.py
